+++ CHANGELOG.md
@@ -28,6 +28,7 @@ mod).
 ### Fixed
 
 - Fixed Eureka+ transforming a card into Distillate instead of Distillate+
+- Fixed the Alchemist not always being targetable by the Mend rest site option in Multiplayer
 
 ## [0.7.1] - 2026-08-11